wlan: add platform board in list check for kalama
BoardConfigWlan.mk got involved twice during ‘lunch’,
and so wlan related modules are added twice, then
there will be 2 items for each in file
'/vendor/lib/modules/modules.load', and it will
try to load each module for 2 times during boot time,
it's incorrect.
To fix this issue, add platform board in list check
before adding the modules:
When BoardConfigWlan.mk is involved at the 1st time,
'device/qcom/common/utils.mk' is not included yet, so
is-board-platform-in-list(which is defined in utils.mk)
is not defined, then
ifeq ($(call is-board-platform-in-list,$(TARGET_BOARD_PLATFORM)),true)
returns false, and no module is added.
At the 2nd time, utils.mk is included, and
ifeq ($(call is-board-platform-in-list,$(TARGET_BOARD_PLATFORM)),true)
returns true, the modules are added.

Change-Id: Ie571c60a4e9d944459d99745d55553d0b1731328
CRs-Fixed: 3346968